test: split cleanup admission cancel cases under 300 lines
Move in-flight hold and post-degrade event omission coverage into
worktree_cleanup_admission_inflight_test.go so neither cancellation
contract file exceeds the AGENTS.md subsystem test-file limit.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
